+++ src/components/Header.tsx
@@ -2,9 +2,9 @@ import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
 
 import { IBook } from '../models/types';
-import SearchBar from './SearchBar';
+import { SearchBar } from './SearchBar';
 
-class Header extends Component<
+export class Header extends Component<
   { onBooksFetched?: (books: IBook[]) => void; hideSearch: boolean },
   object
 > {
@@ -15,6 +15,7 @@ class Header extends Component<
           <div className="header-links">
             <Link to="/">Home</Link>
             <Link to="/about">About</Link>
+            <Link to="/form">Forms</Link>
           </div>
           {!this.props.hideSearch && <SearchBar onBooksFetched={this.props.onBooksFetched} />}
         </header>
@@ -22,5 +23,3 @@ class Header extends Component<
     );
   }
 }
-
-export default Header;
