+++ src/pages/HomePage.tsx
@@ -1,12 +1,12 @@
 import React, { Component } from 'react';
 
-import Header from '../components/Header';
-import Cards from '../components/Cards';
+import { Header } from '../components/Header';
+import { Cards } from '../components/Cards';
 
 import { IBook } from '../models/types';
 import { books } from '../data/data';
 
-class HomePage extends Component<object, { books: IBook[] }> {
+export class HomePage extends Component<object, { books: IBook[] }> {
   constructor(props: object) {
     super(props);
     this.state = { books: books };
@@ -28,5 +28,3 @@ class HomePage extends Component<object, { books: IBook[] }> {
     );
   }
 }
-
-export default HomePage;
