+++ src/pages/NotfoundPage.tsx
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
 
-import Header from '../components/Header';
+import { Header } from '../components/Header';
 
-class NotFoundPage extends Component {
+export class NotFoundPage extends Component {
   render() {
     return (
       <div>
@@ -19,5 +19,3 @@ class NotFoundPage extends Component {
     );
   }
 }
-
-export default NotFoundPage;
